+++ sccoda/util/comp_ana.py
@@ -45,7 +45,7 @@ class CompositionalAnalysis:
         cell_types = data.var.index.to_list()
 
         # Get count data
-        data_matrix = data.X.astype("float32")
+        data_matrix = data.X.astype("float64")
 
         # Build covariate matrix from R-like formula
         covariate_matrix = pt.dmatrix(formula, data.obs)
